requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
Pillow
google-genai
python-multipart
